Initial support of having 3 separate MST trees.
There is now a container object that is keeping 3 copies of the shared model in sync.
It is working by applying the snapshot to each of the tiles using the shared model.
There is no delay so there isn't much possibly of introducing conflicts.

The biggest amount of work was dealing with onInvalidated.
It didn't work well when apply snapshots. I'm pretty sure I can clean up this code and
remove the onInvalidated all together since it is being handled by the autorun now.

Next steps:
- try removing onInvalidated
- add some artificial delay to see how conflicts are handled
- clean up the notes and comments
- try adding undo
- add state saving and loading
- add async loading to work out how to handle incomplete loads that have dependencies