remove onAction listeners, add uuids
this removes the onAction listeners used to sync the shared models from
the tiles to the container. They are replaced by callbacks in the undo recorder
middleware.

It also adds uuids to the actions recorded by the undo recorder.

This isn't complete yet, but the UI works the same as before so it is useful
to commit this intermediate step.

What is left is to remove the autorun from the 2 tile models. These autoruns
don't allow us to track the container action id. They will be replaced with
an action that is called by the middleware and maybe container.  The middleware
will call it when it detects changes in the cached sharedModel. The
container might need to call it when it apply changes to the cached sharedModel