replace autorun based syncing
the new approach allows us to record the containerActionId with changes
caused by the sync. That part of the work isn't done yet though.
This just removes the autorun's.

The new approach is less efficient in some ways and more in others.
It does not use mobx's monitoring of read prepoerties to know when the
sync needs to be re-run. So for example when the name of a node is
changed, there is no need to run the sync. The autorun approach would
not have run in this case. But the new approach will run in this case.
This is because it runs whenever anything is changed in the sharedModel
part of the tree.

However it is more efficient because it is not watching changes to the
tile's part fo the tree. So when a new node is added to the tiles list of
nodes, the sync code will no longer run.