save the undo stack to clipboard
this is a first pass of serializing something
Note that it includes patches and inversePatches so this is great for going forward and backward in time.
But also note this is just the undo stack, so it will include undone actions that could be redone.
It also won't record the history of undo and redoing that a user does.